fix: 삭제하려는 멤버가 클럽 소속이 아닌 경우 예외처리

+++ src/main/java/GDG/whatssue/domain/member/service/ClubMemberManagingService.java
@@ -21,10 +21,14 @@ public class ClubMemberManagingService {
     private final ClubMemberRepository clubMemberRepository;
 
     @Transactional
-    public void deleteClubMember(Long memberId) {
+    public void deleteClubMember(Long clubId, Long memberId) {
 
         ClubMember clubMember = clubMemberRepository.findById(memberId)
             .orElseThrow(() -> new CommonException(ClubMemberErrorCode.EX2100));
+
+        if (clubMember.getClub().getId() != clubId){
+            throw new CommonException(ClubMemberErrorCode.EX2203);
+        }
 
         clubMemberRepository.delete(clubMember);
     }
